// src/types/currencyRates.ts
/* ------------------------------------------------------ type: CurrencyRates ----------------------------------------------------- */

export type CurrencyRates = {
    base: string
    rates: { [key: string]: number }
}


/* -------------------------------------------------------------------------------------------------------------------------------- */